chore: e2e: remove e2e-tests directory
