--- packages/ui/components/combobox/test/e2e-use-cases/lion-combobox-doesn-t-select-any-similar-options-after-using-delete-when-requireOptionMatch-is-false.js
@@ -1,19 +1,0 @@
-import { html, render } from 'lit';
-import { Required } from '@lion/ui/form-core.js';
-import '@lion/ui/define/lion-combobox.js';
-import '@lion/ui/define/lion-option.js';
-import { loadDefaultFeedbackMessages } from '@lion/ui/validate-messages.js';
-
-loadDefaultFeedbackMessages();
-
-const template = () =>
-  html` 
-    <lion-combobox name="foo" .validators=${[new Required()]}>
-      <lion-option .choiceValue="${'Artichoke'}">Artichoke</lion-option>
-      <lion-option .choiceValue="${'Chard'}">Chard</lion-option>
-      <lion-option .choiceValue="${'Chicory'}">Chicory</lion-option>
-      <lion-option .choiceValue="${'Victoria Plum'}">Victoria Plum</lion-option>
-    </lion-combobox>
-  `;
-
-render(template(), document.querySelector('e2e-root'));
